requirements.txt
google-genai==1.7.0
faiss
python-dotenv
numpy
streamlit
fastapi
uvicorn
langchain_google_genai
langchain
langchain_community
langchain_core